Separate http library from server
