+++ container/shim/start.sh
@@ -21,4 +21,4 @@ else
 fi
 
 nginx -g "daemon off;" &
-exec node src/index.js
+exec node bin/shim.js
